Add task metadata to /apps endpoint (#118)
* Export TaskInfo from the mesos agent collector

We will reference TaskInfo in a map of container IDs to task info in a
later commit.

* Add a struct for container->task relationships

ContainerTaskRels has a single member, a map of container IDs to
TaskInfo references. It has a getter and a setter for easy access
threadsafe access to the relationship map, and an update method which
populates it from agentState.

* Return task_id and task_name in /app output

The new ContainerTaskRels argument providers access to the task metadata
for each container.

* Add tests for extracting a record with task data

We add a new mock object with appropriate metadata and supply it to
avroRecord.extract(), checking that the task IDs in the
ContainerTaskRels are the same ones that correspond to the container
ID.

* Add tests for executor ID and framework ID

This is orthogonal to the intent of this PR, however it's a small change
and seemed like a quick win.

* Refer to ContainerTaskRels in Framework collector

This reference can be provided to the extract method from the framework
collector.

* Pass ref to ContainerTaskRels into transform

* Update task mapping from agent collector

We update the task mapping after every new fetch from agent state.

* Pass reference to mapping into framework collector

Now the framework collector has access to the ContainerTaskRels object
which belongs to the mesos agent collector.

* Clarify arg name

* Defer unlock for consistency

* Temporarily skip golint in build

In scripts/test.sh we naively run go get to update golint. A new rule
has just been added to golint (https://github.com/golang/lint/pull/319)
and I don't want to complicate my current PR (#118).

I will revert this commit, fix any outstanding nil-returns and consider
pinning golint in a future PR.

* Create new TaskInfo object to avoid memory issues

* Create ContainerTaskRels in dcos-metrics.go

Rather than initiating ContainerTaskRels in the framework collector, we
now initiate it in dcos-metrics.go and pass a reference into each
collector instead.

* Initiate ContainerTaskRels with a utility method

* Remove unnecessary nil check